Pilot spike to prototype basic structure and functionality.
This commit builds out the basic Norris data structures (Instance,
Project, Activity, Issue) and uses them to compose a "flow" document
for a particular issue type showing a snapshot of inventories and
recent transitions for those issues.

Invocation through the command-line utility looks like:

    bin/nrs --config config/norris.json --project UAR --activity Task --flow

That assumes that you've copied the "norris.json.template" file to
"norris.json", and replaced the placeholders for Jira URL, user, and
password with real values.

There's a lot of cleanup and refactoring left to be done, but
this gets the basic concepts on the board for further refinement.