Second pass to refine basic structure and functionality.
This commit breaks down the original "nrs" script into something
much more like the toolkit approach favored by UNIX.  Functionality
is split into multiple smaller scripts, and much of the library has
been simplified.  All of the scripting is designed to be pipelined,
e.g. something like this:

    bin/nrs-search -q '( issuekey in ( UAR-10, UAR-11, UAR-12 ) )'
    | bin/nrs-extract 
    | bin/nrs-i2txt
    
This does a query, extracts the issues for the returned keys, and then
converts them to a textual representation.

Note also that the location of the Norris config file can either be
specified with a command line argument, or can now be specified by
setting the "NRSCONFIG" environment variable (so that the pipelines
can be a little more shorter and readable).

Next step is to add some additional utility scripts for different
kinds of manipulation or output, and see what kind of growth that
generates in the library code.